+++ Calculator/src/main/java/com/aparna/learning/java/Calculator/CalculatorInterface.java
@@ -1,8 +1,8 @@
 package com.aparna.learning.java.Calculator;
 
 public interface CalculatorInterface {
-	public int add(int a, int b);
-	public int sub(int a,int b);
-	public int mul(int a,int b);
-	public int div(int a,int b) throws Exception;
+	public double add(double a, double b);
+	public double sub(double a,double b);
+	public double mul(double a,double b);
+	public double div(double a, double b) throws Exception;
 }
